[pstoken.c] long to int fixes

+++ c/public/lib/api/dynarr.h
@@ -71,7 +71,7 @@ typedef struct dnaCtx_ *dnaCtx;
         dnaCtx ctx;                                    \
         type *array;                                   \
         long cnt;                                      \
-        long size;                                     \
+        int32_t size;                                  \
         long incr;                                     \
         void (*func)(void *ctx, long cnt, type *base); \
     } da
